Clean up custom build logic for generating API incubation reports
- Use the latest (and not deprecated) version of Worker API
- Move all aggregation logic to buildSrc
- Use proper variants and dependencies to aggregate reports
