+++ buildSrc/subprojects/buildquality/src/main/kotlin/gradlebuild/incubation/action/IncubatingApiReportAggregationParameter.kt
@@ -13,10 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-plugins {
-    id("gradlebuild.incubation-report-aggregation")
-}
 
-dependencies {
-    reports(platform(project(":distributions-full")))
+package gradlebuild.incubation.action
+
+import org.gradle.api.file.ConfigurableFileCollection
+import org.gradle.api.file.RegularFileProperty
+import org.gradle.workers.WorkParameters
+
+
+interface IncubatingApiReportAggregationParameter : WorkParameters {
+    val reports: ConfigurableFileCollection
+    val htmlReportFile: RegularFileProperty
 }
